# Optimize concat Presto function with consecutive constant inputs