--- a/velox/functions/prestosql/StringFunctions.cpp
+++ b/velox/functions/prestosql/StringFunctions.cpp
@@ -143,6 +143,53 @@ class UpperLowerTemplateFunction : public exec::VectorFunction {
  * */
 class ConcatFunction : public exec::VectorFunction {
  public:
+  ConcatFunction(
+      const std::string& /* name */,
+      const std::vector<exec::VectorFunctionArg>& inputArgs) {
+    auto numArgs = inputArgs.size();
+
+    // Save constant values to constantStrings_.
+    // Identify and combine consecutive constant inputs.
+    argMapping_.reserve(numArgs);
+    constantStrings_.reserve(numArgs);
+
+    for (auto i = 0; i < numArgs; ++i) {
+      argMapping_.push_back(i);
+
+      const auto& arg = inputArgs[i];
+      if (arg.constantValue) {
+        std::string value = arg.constantValue->as<ConstantVector<StringView>>()
+                                ->valueAt(0)
+                                .str();
+
+        column_index_t j = i + 1;
+        for (; j < inputArgs.size(); ++j) {
+          if (!inputArgs[j].constantValue) {
+            break;
+          }
+
+          value += inputArgs[j]
+                       .constantValue->as<ConstantVector<StringView>>()
+                       ->valueAt(0)
+                       .str();
+        }
+
+        constantStrings_.push_back(std::string(value.data(), value.size()));
+
+        i = j - 1;
+      } else {
+        constantStrings_.push_back(std::string());
+      }
+    }
+
+    // Create StringViews for constant strings.
+    constantStringViews_.reserve(numArgs);
+    for (const auto& constantString : constantStrings_) {
+      constantStringViews_.push_back(
+          StringView(constantString.data(), constantString.size()));
+    }
+  }
+
   bool propagateStringEncodingFromAllInputs() const override {
     return true;
   }
@@ -156,14 +203,31 @@ class ConcatFunction : public exec::VectorFunction {
     context.ensureWritable(rows, VARCHAR(), result);
     auto flatResult = result->asFlatVector<StringView>();
 
-    exec::DecodedArgs decodedArgs(rows, args, context);
+    auto numArgs = argMapping_.size();
+
+    std::vector<exec::LocalDecodedVector> decodedArgs;
+    decodedArgs.reserve(numArgs);
+
+    for (auto i = 0; i < numArgs; ++i) {
+      auto index = argMapping_[i];
+      if (constantStringViews_[i].empty()) {
+        decodedArgs.emplace_back(context, *args[index], rows);
+      } else {
+        // Do not decode constant inputs.
+        decodedArgs.emplace_back(context);
+      }
+    }
 
     // Calculate the combined size of the result strings.
     size_t totalResultBytes = 0;
     rows.applyToSelected([&](int row) {
-      for (int i = 0; i < args.size(); i++) {
-        auto value = decodedArgs.at(i)->valueAt<StringView>(row);
-        totalResultBytes += value.size();
+      for (int i = 0; i < numArgs; i++) {
+        if (constantStringViews_[i].empty()) {
+          auto value = decodedArgs[i]->valueAt<StringView>(row);
+          totalResultBytes += value.size();
+        } else {
+          totalResultBytes += constantStringViews_[i].size();
+        }
       }
     });
 
@@ -176,8 +240,13 @@ class ConcatFunction : public exec::VectorFunction {
       const char* start = rawBuffer + offset;
 
       size_t combinedSize = 0;
-      for (int i = 0; i < args.size(); i++) {
-        auto value = decodedArgs.at(i)->valueAt<StringView>(row);
+      for (int i = 0; i < numArgs; i++) {
+        StringView value;
+        if (constantStringViews_[i].empty()) {
+          value = decodedArgs[i]->valueAt<StringView>(row);
+        } else {
+          value = constantStringViews_[i];
+        }
         auto size = value.size();
         if (size > 0) {
           memcpy(rawBuffer + offset, value.data(), size);
@@ -202,6 +271,11 @@ class ConcatFunction : public exec::VectorFunction {
   static exec::VectorFunctionMetadata metadata() {
     return {true /* supportsFlattening */};
   }
+
+ private:
+  std::vector<column_index_t> argMapping_;
+  std::vector<std::string> constantStrings_;
+  std::vector<StringView> constantStringViews_;
 };
 
 /**
@@ -495,11 +569,13 @@ VELOX_DECLARE_VECTOR_FUNCTION(
     UpperLowerTemplateFunction<true /*isLower*/>::signatures(),
     std::make_unique<UpperLowerTemplateFunction<true /*isLower*/>>());
 
-VELOX_DECLARE_VECTOR_FUNCTION_WITH_METADATA(
+VELOX_DECLARE_STATEFUL_VECTOR_FUNCTION_WITH_METADATA(
     udf_concat,
     ConcatFunction::signatures(),
     ConcatFunction::metadata(),
-    std::make_unique<ConcatFunction>());
+    [](const auto& name, const auto& inputs) {
+      return std::make_unique<ConcatFunction>(name, inputs);
+    });
 
 VELOX_DECLARE_VECTOR_FUNCTION(
     udf_strpos,