Summary:
concat(c0, ',', c1, ',', '100', ',', '200) was 2x slower than concat
(c0, ',', c1, ',100,200'). This change is to optimize concat function
implementation to detect consecutive constant inputs and combine these into
single value. The function is now a stateful vector function.

Before:

```
============================================================================
[...]stosql/benchmarks/ConcatBenchmark.cpp     relative  time/iter   iters/s
============================================================================
basic                                                     176.31ns     5.67M
flatten                                                   172.56ns     5.79M
flattenAndConstantFold                                     99.07ns    10.09M
```

After:

```
============================================================================
[...]stosql/benchmarks/ConcatBenchmark.cpp     relative  time/iter   iters/s
============================================================================
basic                                                      70.25ns    14.24M
flatten                                                    71.88ns    13.91M
flattenAndConstantFold                                     71.83ns    13.92M
```

Differential Revision: D39622939